gdataset: correctly handle NULL string in g_datalist_get_data()
All code that adds an entry to GData calls down to datalist_append().
No caller would never pass a zero key_id. So a GData will never contain
the zero GQuark.

g_datalist_get_data() allows to lookup the keys by string. Looking up
by string, only makes sense if the string is a valid GQuark. And the
NULL string is not a valid GQuark (or, in another interpretation is to
say that the GQuark of NULL is zero). In any case, the GData does not
contain the zero quark and it should never find a NULL string there.

Note however that you can add invalid (non-zero) GQuarks to GData. That
is because we avoid the overhead of checking that the non-zero key_is is
infact a valid GQuark. Thus, if you called

  g_datalist_get_data(&data, NULL)

the check

  if (g_strcmp0 (g_quark_to_string (data_elt->key), key) == 0)

would have found entries where the key is not a valid GQuark.

Fix that. Looking up a NULL string should never finds an entry.

Fixes: f6a9d0479622 ('GDataSet: silently accept NULL/0 as keys')